Made DummyInterpreter
